refactor(exec): extract session handle event pipeline
