+++ src/JKToolKit.CodexSDK/Exec/CodexSessionHandle.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using JKToolKit.CodexSDK.Abstractions;
+using JKToolKit.CodexSDK.Exec.Internal;
 using JKToolKit.CodexSDK.Exec.Notifications;
 using JKToolKit.CodexSDK.Exec.Protocol;
 using Microsoft.Extensions.Logging;
@@ -110,10 +111,20 @@ public sealed class CodexSessionHandle : ICodexSessionHandle
 
         var lineStream = _tailer.TailAsync(Info.LogPath, effectiveOptions, pipelineCts.Token);
         var parsedStream = _parser.ParseAsync(lineStream, pipelineCts.Token);
-        var filteredStream = ApplyTimestampFilter(parsedStream, effectiveOptions, pipelineCts.Token);
+
+        var pipeline = new CodexSessionHandleEventPipeline(
+            Info,
+            _process,
+            _processLauncher,
+            _processExitTimeout,
+            _logger,
+            notifyExit: NotifyExitSafe,
+            tryStartIdleTermination: TryStartIdleTermination);
+
+        var filteredStream = pipeline.ApplyTimestampFilter(parsedStream, effectiveOptions, pipelineCts.Token);
 
         var finalStream = (_idleTimeout.HasValue && shouldFollow && _process != null && _processLauncher != null)
-            ? ApplyIdleTimeout(filteredStream, _idleTimeout.Value, pipelineCts, cancellationToken)
+            ? pipeline.ApplyIdleTimeout(filteredStream, _idleTimeout.Value, pipelineCts, cancellationToken)
             : filteredStream;
 
         await foreach (var evt in finalStream)
@@ -304,6 +315,9 @@ public sealed class CodexSessionHandle : ICodexSessionHandle
         }
     }
 
+    private bool TryStartIdleTermination() =>
+        Interlocked.Exchange(ref _idleTerminationStarted, 1) == 0;
+
     /// <inheritdoc />
     public IDisposable OnExit(Action<int> callback)
     {
@@ -366,173 +380,5 @@ public sealed class CodexSessionHandle : ICodexSessionHandle
         }
     }
 
-    private async IAsyncEnumerable<CodexEvent> ApplyIdleTimeout(
-        IAsyncEnumerable<CodexEvent> events,
-        TimeSpan idleTimeout,
-        CancellationTokenSource pipelineCts,
-        [EnumeratorCancellation] CancellationToken userCancellationToken)
-    {
-        if (idleTimeout <= TimeSpan.Zero)
-        {
-            await foreach (var evt in events.WithCancellation(userCancellationToken))
-            {
-                yield return evt;
-            }
 
-            yield break;
-        }
-
-        long lastMessageTicks = 0;
-        int idleTriggered = 0;
-        using var monitorCts = CancellationTokenSource.CreateLinkedTokenSource(userCancellationToken, pipelineCts.Token);
-        var monitorTask = MonitorIdleAsync(
-            idleTimeout,
-            () => Interlocked.Read(ref lastMessageTicks),
-            pipelineCts,
-            monitorCts.Token,
-            () => Interlocked.Exchange(ref idleTriggered, 1));
-
-        await using var enumerator = events.WithCancellation(monitorCts.Token).GetAsyncEnumerator();
-
-        try
-        {
-            while (true)
-            {
-                bool hasNext;
-                try
-                {
-                    hasNext = await enumerator.MoveNextAsync();
-                }
-                catch (OperationCanceledException) when (Volatile.Read(ref idleTriggered) == 1 && !userCancellationToken.IsCancellationRequested)
-                {
-                    yield break;
-                }
-
-                if (!hasNext)
-                {
-                    yield break;
-                }
-
-                Interlocked.Exchange(ref lastMessageTicks, DateTimeOffset.UtcNow.UtcTicks);
-                yield return enumerator.Current;
-            }
-        }
-        finally
-        {
-            monitorCts.Cancel();
-            try
-            {
-                await monitorTask.ConfigureAwait(false);
-            }
-            catch (OperationCanceledException)
-            {
-                // Expected if cancellation/token triggered
-            }
-            catch (Exception ex)
-            {
-                _logger.LogDebug(ex, "Idle timeout monitor encountered an error for session {SessionId}", Info.Id);
-            }
-        }
-    }
-
-    private async Task MonitorIdleAsync(
-        TimeSpan idleTimeout,
-        Func<long> lastMessageTicksProvider,
-        CancellationTokenSource pipelineCts,
-        CancellationToken cancellationToken,
-        Action onIdleTriggered)
-    {
-        while (!cancellationToken.IsCancellationRequested)
-        {
-            var lastTicks = lastMessageTicksProvider();
-            if (lastTicks == 0)
-            {
-                try
-                {
-                    await Task.Delay(TimeSpan.FromMilliseconds(100), cancellationToken).ConfigureAwait(false);
-                }
-                catch (OperationCanceledException)
-                {
-                    break;
-                }
-
-                continue;
-            }
-
-            var last = new DateTimeOffset(lastTicks, TimeSpan.Zero);
-            var elapsed = DateTimeOffset.UtcNow - last;
-            var remaining = idleTimeout - elapsed;
-
-            if (remaining <= TimeSpan.Zero)
-            {
-                _logger.LogInformation(
-                    "Idle timeout of {IdleTimeout} elapsed for session {SessionId}; terminating Codex process.",
-                    idleTimeout,
-                    Info.Id);
-
-                onIdleTriggered();
-                await TriggerIdleTerminationAsync().ConfigureAwait(false);
-                pipelineCts.Cancel();
-                return;
-            }
-
-            try
-            {
-                await Task.Delay(remaining, cancellationToken).ConfigureAwait(false);
-            }
-            catch (OperationCanceledException)
-            {
-                // cancellation or pipeline shut down
-                break;
-            }
-        }
-    }
-
-    private async Task TriggerIdleTerminationAsync()
-    {
-        if (_process == null || _processLauncher == null)
-        {
-            return;
-        }
-
-        if (Interlocked.Exchange(ref _idleTerminationStarted, 1) != 0)
-        {
-            return;
-        }
-
-        try
-        {
-            var exitCode = await _processLauncher
-                .TerminateProcessAsync(_process, _processExitTimeout, CancellationToken.None)
-                .ConfigureAwait(false);
-
-            NotifyExitSafe(exitCode, SessionExitReason.Timeout);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(
-                ex,
-                "Failed to terminate Codex process for session {SessionId} after idle timeout.",
-                Info.Id);
-        }
-    }
-
-    private async IAsyncEnumerable<CodexEvent> ApplyTimestampFilter(
-        IAsyncEnumerable<CodexEvent> events,
-        EventStreamOptions options,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        await foreach (var evt in events.WithCancellation(cancellationToken))
-        {
-            if (options.AfterTimestamp.HasValue)
-            {
-                if (evt.Timestamp <= options.AfterTimestamp.Value)
-                {
-                    continue;
-                }
-            }
-
-            yield return evt;
-        }
-    }
 }
